+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Macro `bydi-match-expansion` now has its own `ert-explainer` to show
+  the mismatch between actual and expected expansion.
+
 ## [v0.2.0]
 
 ### Added